move auto_complete tests to another file, add zsh&bash variants for some cases
